+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridge.kt
@@ -2,12 +2,10 @@ package com.rakuten.tech.mobile.miniapp.js
 
 import android.app.Activity
 import android.content.Intent
-import android.os.Build
 import android.webkit.JavascriptInterface
 import androidx.annotation.VisibleForTesting
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import com.rakuten.tech.mobile.miniapp.BuildConfig
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
 import com.rakuten.tech.mobile.miniapp.R
 import com.rakuten.tech.mobile.miniapp.CustomPermissionsNotImplementedException
@@ -77,20 +75,13 @@ open class MiniAppMessageBridge {
     @VisibleForTesting
     internal fun createBridgeExecutor(webViewListener: WebViewListener) = MiniAppBridgeExecutor(webViewListener)
 
-    /**
-     * Get provided id of mini app for any purpose.
-     * You can also throw an [Exception] from this method to pass an error message to the mini app.
-     */
-    @Deprecated("This function has been deprecated.",
-            ReplaceWith("getUniqueId(onSuccess: (uniqueId: String) -> Unit," +
-                            "onError: (message: String) -> Unit)"
-            )
-    )
-    open fun getUniqueId(): String = throw MiniAppSdkException(ErrorBridgeMessage.NO_IMPL)
-
     /** Get provided id of mini app for any purpose. **/
-    open fun getUniqueId(onSuccess: (uniqueId: String) -> Unit, onError: (message: String) -> Unit) =
-        onSuccess(getUniqueId())
+    open fun getUniqueId(
+        onSuccess: (uniqueId: String) -> Unit,
+        onError: (message: String) -> Unit
+    ) {
+        throw MiniAppSdkException(ErrorBridgeMessage.NO_IMPL)
+    }
 
     /** Post device permission request from external. **/
     open fun requestDevicePermission(
@@ -148,14 +139,7 @@ open class MiniAppMessageBridge {
         if (!locale.isValidLocale())
             locale = ""
 
-        val hostEnvironmentInfo = HostEnvironmentInfo(
-                platformVersion = Build.VERSION.RELEASE,
-                hostVersion = activity.packageManager.getPackageInfo(
-                        activity.packageName, 0
-                ).versionName,
-                sdkVersion = BuildConfig.VERSION_NAME,
-                hostLocale = locale
-        )
+        val hostEnvironmentInfo = HostEnvironmentInfo(activity = activity, hostLocale = locale)
 
         onSuccess.invoke(hostEnvironmentInfo)
     }