+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/analytics/rat_wrapper/RATFragment.kt
@@ -4,7 +4,10 @@ import androidx.fragment.app.Fragment
 import com.rakuten.tech.mobile.testapp.analytics.DemoAppAnalytics
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 
-abstract class RATFragment: Fragment(), RatComponent {
+/**
+ * This is a custom Fragment to handle rat analytics.
+ */
+abstract class RATFragment : Fragment(), RatComponent {
     override fun onResume() {
         DemoAppAnalytics.init(AppSettings.instance.projectId).sendAnalytics(
             RATEvent(
@@ -16,5 +19,3 @@ abstract class RATFragment: Fragment(), RatComponent {
         super.onResume()
     }
 }
-
-